feat: currentMeeting store create
